@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pg-alchemy-kit"
-version = "0.8.3"
+version = "0.9.0-beta"
 description = "A simple wrapper around sqlalchemy to make it easier to use with postgresql"
 authors = ["jsaied99 <jsaied99@gmail.com>"]
 readme = "README.md"
@@ -8,9 +8,9 @@ repository = "https://github.com/jsaied99/pg-alchemy-kit"
 packages = [{ include = "pg_alchemy_kit" }]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.11"
 psycopg2-binary = "2.9.3"
-sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
+sqlalchemy = { extras = ["asyncio"], version = "^2.0.27" }
 pandas = "^2.2.0"
 pyarrow = "^15.0.0"
 asyncpg = "^0.29.0"
@@ -19,3 +19,16 @@ asyncpg = "^0.29.0"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+
+[tool.basedpyright]
+exclude = ["**/__pycache__"]
+defineConstant = { DEBUG = true }
+
+reportMissingImports = true
+reportMissingTypeStubs = false
+reportAny = false
+reportUninitializedInstanceVariable = false
+
+pythonVersion = "3.12"
+pythonPlatform = "Linux"
